pyproject.toml
[project]
name = "universal-bs4"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]
